Add time breakdown into suite test report and changelog

## Motivation

(Write your motivation for proposed changes here.)

### Have you read the [Contributing Guidelines on pull requests](https://github.com/diem/diem/blob/master/CONTRIBUTING.md#pull-requests)?

(Write your answer here.)

## Test Plan
./scripts/cti --pr 7735 --suite perf

Before:
```
all up : 926 TPS, 4902 ms latency, 5600 ms p99 latency, no expired txns
10% down : 894 TPS, 4563 ms latency, 7650 ms p99 latency, no expired txns
3 Region Simulation : 733 TPS, 6209 ms latency, 7850 ms p99 latency, no expired txns
fixed tps 10 : 10 TPS, 924 ms latency, 1200 ms p99 latency, no expired txns
```

After:
```
Test runner setup time spent 298 secs
all up : 910 TPS, 4994 ms latency, 5700 ms p99 latency, no expired txns
Time spent 274 secs
10% down : 887 TPS, 4610 ms latency, 6500 ms p99 latency, no expired txns
Time spent 419 secs
3 Region Simulation : 724 TPS, 6305 ms latency, 7800 ms p99 latency, no expired txns
Time spent 267 secs
fixed tps 10 : 10 TPS, 913 ms latency, 1000 ms p99 latency, no expired txns
Time spent 248 secs
```

(Share your test plan here. If you changed code, please provide us with clear instructions for verifying that your changes work.)

## Related PRs

(If this PR adds or changes functionality, please take some time to update the docs at https://github.com/diem/diem/tree/master/developers.diem.com, and link to your PR here.)